Add Coinbase api as back up for exchange prices
