+++ src/lib/constants.ts
@@ -6,9 +6,11 @@ export const MODE = import.meta.env.MODE
 
 export const lnsocketProxy = 'wss://lnsocket.clams.tech'
 
-export const BITCOIN_EXCHANGE_RATE_ENDPOINT = `https://api.coingecko.com/api/v3/simple/price?ids=bitcoin&vs_currencies=${Object.keys(
+export const COIN_GECKO_PRICE_ENDPOINT = `https://api.coingecko.com/api/v3/simple/price?ids=bitcoin&vs_currencies=${Object.keys(
 	FiatDenomination
 ).join(',')}`
+
+export const COINBASE_PRICE_ENDPOINT = 'https://api.coinbase.com/v2/exchange-rates?currency=BTC'
 
 export const SEC_IN_MS = 1000
 export const MIN_IN_MS = 60 * 1000
