+++ src/routes/receive.svelte
@@ -6,8 +6,8 @@
 	import Amount from '$lib/components/Amount.svelte'
 	import Description from '$lib/components/Description.svelte'
 	import Summary from '$lib/components/Summary.svelte'
-	import { addPayment, settings$, waitForAndUpdatePayment } from '$lib/streams'
-	import { SvelteSubject } from '$lib/utils'
+	import { settings$ } from '$lib/streams'
+	import { SvelteSubject, waitForAndUpdatePayment, addPayment } from '$lib/utils'
 	import { convertValue } from '$lib/conversion'
 	import { goto } from '$app/navigation'
 	import { BitcoinDenomination } from '$lib/types'
