+++ src/routes/send.svelte
@@ -8,9 +8,9 @@
 	import Summary from '$lib/components/Summary.svelte'
 	import Slide from '$lib/elements/Slide.svelte'
 	import { BitcoinDenomination, type PaymentType } from '$lib/types'
-	import { SvelteSubject } from '$lib/utils'
+	import { SvelteSubject, addPayment } from '$lib/utils'
 	import { convertValue } from '$lib/conversion'
-	import { addPayment, settings$ } from '$lib/streams'
+	import { settings$ } from '$lib/streams'
 	import Amount from '$lib/components/Amount.svelte'
 	import Description from '$lib/components/Description.svelte'
 	import { t } from '$lib/i18n/translations'
