+++ src/routes/welcome.svelte
@@ -24,9 +24,9 @@
 	import { t } from '$lib/i18n/translations'
 	import TextInput from '$lib/elements/TextInput.svelte'
 	import Button from '$lib/elements/Button.svelte'
-	import { credentials$, updateCredentials } from '$lib/streams'
+	import { credentials$ } from '$lib/streams'
 	import { coreLightning, type Socket } from '$lib/backends'
-	import { truncateValue, validateConnectionString } from '$lib/utils'
+	import { truncateValue, validateConnectionString, updateCredentials } from '$lib/utils'
 	import Check from '$lib/icons/Check.svelte'
 	import Close from '$lib/icons/Close.svelte'
 	import Arrow from '$lib/icons/Arrow.svelte'
